Only login and presentation page is accessible without begin authenticated.

+++ src/server/routes/hybrids.js
@@ -1,5 +1,6 @@
 import { connectDB } from '../config/database';
 import { v4 as uuid } from 'uuid';
+import isAuth from './isAuth';
 
 export const hybrids = app => {
   const addNewHybrid = async hybrid => {
@@ -8,30 +9,38 @@ export const hybrids = app => {
     await collection.insertOne(hybrid);
     return hybrid;
   };
-  app.post('/hybrid/new', async (req, res) => {
-    const newHybrid = await addNewHybrid({ id: uuid(), ...req.body });
+  app.post('/hybrid/new', isAuth, async (req, res) => {
+    const newHybrid = await addNewHybrid({
+      id: uuid(),
+      user: req.session.passport.user,
+      ...req.body,
+    });
     res.status(200).json(newHybrid);
   });
+
+  const checkHybridOwner = async (hybridId, user) => {
+    const db = await connectDB();
+    const collection = db.collection('hybrids');
+    const hybrid = collection.findOne({ id: hybridId });
+    return hybrid.user === user.id;
+  };
 
   const deleteHybrid = async hybridId => {
     const db = await connectDB();
     const collection = db.collection('hybrids');
     await collection.deleteOne({ id: hybridId });
   };
-  app.delete('/hybrid/:id', async (req, res) => {
-    await deleteHybrid(req.params.id);
-    res.status(200).send();
-  });
-
-  const saveHybrid = async hybrid => {
-    const db = await connectDB();
-    const collection = db.collection('hybrids');
-    await collection.save(hybrid);
-  };
-  app.post('/hybrid/save', async (req, res) => {
-    const hybrid = req.body.hybrid;
-    await saveHybrid(hybrid);
-    res.status(200).send();
+  app.delete('/hybrid/:id', isAuth, async (req, res) => {
+    const isOwner = await checkHybridOwner(
+      req.params.id,
+      req.session.passport.user
+    );
+    if (isOwner) {
+      await deleteHybrid(req.params.id);
+      res.status(200).send();
+    } else {
+      res.status(401).send();
+    }
   });
 
   const updateHybrid = async hybrid => {
@@ -55,8 +64,16 @@ export const hybrids = app => {
       await collection.updateOne({ id }, { $set: { grid } });
     }
   };
-  app.post('/hybrid/update', async (req, res) => {
-    await updateHybrid(req.body);
-    res.status(200).send();
+  app.post('/hybrid/update', isAuth, async (req, res) => {
+    const isOwner = await checkHybridOwner(
+      req.body.id,
+      req.session.passport.user
+    );
+    if (isOwner) {
+      await updateHybrid(req.body);
+      res.status(200).send();
+    } else {
+      res.status(401).send();
+    }
   });
 };
